# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Clientes",
        "description": "Mostando la Lista de URI's de mi API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://apiparaprincipiantesv10.test:8081"
        }
    ],
    "paths": {
        "/api/clientes": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Mostrar Lista de Clientes",
                "operationId": "de55d50d1b708d9e5c63ee3f8fe8f54a",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombres": {
                                                        "type": "string",
                                                        "example": "Aderson Felix"
                                                    },
                                                    "apellidos": {
                                                        "type": "string",
                                                        "example": "Jara Lazaro"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Registrar la información de un Cliente",
                "operationId": "4ac6cdfa8d6c188fa2cb4f3263d4318a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombres": {
                                                "type": "string"
                                            },
                                            "apellidos": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombres": "Aderson Felix",
                                    "apellidos": "Jara Lazaro"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombres": {
                                            "type": "string",
                                            "example": "Aderson Felix"
                                        },
                                        "apellidos": {
                                            "type": "string",
                                            "example": "Jara Lazaro"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Mostrar la información de un cliente",
                "operationId": "0d3826d79390929601421b5810b61231",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombres": {
                                            "type": "string",
                                            "example": "Aderson Felix"
                                        },
                                        "apellidos": {
                                            "type": "string",
                                            "example": "Jara Lazaro"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Actualizar la información de un Cliente",
                "operationId": "74956b09f3ea120b5dd531c73776ca11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombres": {
                                                "type": "string"
                                            },
                                            "apellidos": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombres": "Aderson Felix Editado",
                                    "apellidos": "Jara Lazaro Editado"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombres": {
                                            "type": "string",
                                            "example": "Aderson Felix Editado"
                                        },
                                        "apellidos": {
                                            "type": "string",
                                            "example": "Jara Lazaro Editado"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Eliminar la información de un cliente",
                "operationId": "d15e304c609752c9f409ab81072fb52e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT"
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "No se pudo realizar correctamente la operación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}